backend/requirements.txt
jsonschema>=4.22
fastapi
uvicorn[standard]
python-dotenv
httpx
geopy>=2.4
polyline
lightning
openai-agents>=0.0.14